Week 9 Lab - Various Powershell Programming
This week I created three separate Powershell programs that each do different things. One copies the list of running services and programs on a system to two separate files, another opens and closes the calculator app utilizing its process name and ID, and the final one shows the DHCP and DNS servers of a systems adapters.