Week 10 - Array Filter Program
This lab we needed to write a program that allowed users to filter installed services by 3 different values for their status. This was "Stopped", "Running", and "all". I utilized an array with those three values and used a series of if and elseif statements to run through the possibilities and prevent mistakes.